+++ frontend/src/features/collections/helpers/injectRwpStyles.ts
@@ -1,0 +1,19 @@
+import type { ReplayWebPage } from "replaywebpage";
+
+import replayStylesheet from "@/features/collections/styles/replay.stylesheet.css";
+
+export function injectRwpStyles(replayEmbed: ReplayWebPage | null | undefined) {
+  if (!replayEmbed) {
+    console.debug("no replayEmbed");
+    return;
+  }
+
+  const iframeDoc =
+    replayEmbed.shadowRoot?.querySelector("iframe")?.contentDocument;
+
+  if (iframeDoc) {
+    const style = iframeDoc.createElement("style");
+    style.textContent = replayStylesheet;
+    iframeDoc.head.appendChild(style);
+  }
+}
