+++ frontend/src/pages/org/collection-detail/collection-detail.ts
@@ -36,9 +36,9 @@ import type { BtrixRequestOrgUpdate } from "@/events/btrix-request-org-update";
 import type { CollectionPageHeader } from "@/features/collections/collection-page-header";
 import { DEFAULT_THUMBNAIL } from "@/features/collections/collection-thumbnail";
 import { collectionRwpContext } from "@/features/collections/context/collection-rwp";
+import { injectRwpStyles } from "@/features/collections/helpers/injectRwpStyles";
 import { collectionShareLink } from "@/features/collections/helpers/share-link";
 import { SelectCollectionAccess } from "@/features/collections/select-collection-access";
-import replayStylesheet from "@/features/collections/styles/replay.stylesheet.css";
 import { createIndexDialog } from "@/features/collections/templates/create-index-dialog";
 import { deleteIndexDialog } from "@/features/collections/templates/delete-index-dialog";
 import { purgeIndexDialog } from "@/features/collections/templates/purge-index-dialog";
@@ -1393,7 +1393,7 @@ export class CollectionDetail extends BtrixElement {
         @rwp-url-change=${(e: RwpUrlChangeEvent) => {
           if (!this.replayEmbed) {
             this.replayEmbed = e.currentTarget as ReplayWebPage;
-            this.injectRwpStyles();
+            injectRwpStyles(this.replayEmbed);
           }
           if (!this.isRwpLoaded) {
             this.isRwpLoaded = true;
@@ -1417,22 +1417,6 @@ export class CollectionDetail extends BtrixElement {
       ></replay-web-page>
     `;
   };
-
-  private injectRwpStyles() {
-    if (!this.replayEmbed) {
-      console.debug("no this.replayEmbed");
-      return;
-    }
-
-    const iframeDoc =
-      this.replayEmbed.shadowRoot?.querySelector("iframe")?.contentDocument;
-
-    if (iframeDoc) {
-      const style = iframeDoc.createElement("style");
-      style.textContent = replayStylesheet;
-      iframeDoc.head.appendChild(style);
-    }
-  }
 
   private readonly renderSpinner = () => html`
     <div
